notification-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
aio-pika==9.4.0
python-jose[cryptography]==3.3.0
prometheus-fastapi-instrumentator==6.1.0
python-multipart==0.0.9
websockets==12.0
